requirements.txt
langchain-groq
langchain
python-dotenv
langchain-community
langchain-core
streamlit
langchain_openai
faiss-cpu
pypdf
fastapi
uvicorn
langchain-openai
jinja2
requests
PyPDF2
tiktoken
python-multipart

-e .